typescript(vx-demo): begin Drag-i TypeScript rewrite, add next.config.js for TS errors
