+++ packages/vx-demo/src/next.config.js
@@ -1,0 +1,7 @@
+// eslint-disable-next-line no-undef
+module.exports = {
+  typescript: {
+    ignoreDevErrors: true,
+    ignoreBuildErrors: true,
+  },
+};
